requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Faker==12.0.1
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
mixer==7.2.2
packaging==21.3
promise==2.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
